+++ src/metric/src/Listener/OnMetricFactoryReady.php
@@ -18,10 +18,12 @@ use Hyperf\Coordinator\Timer;
 use Hyperf\Engine\Coroutine as Co;
 use Hyperf\Event\Contract\ListenerInterface;
 use Hyperf\Metric\Contract\MetricFactoryInterface;
+use Hyperf\Metric\CoroutineServerStats;
 use Hyperf\Metric\Event\MetricFactoryReady;
 use Hyperf\Metric\MetricFactoryPicker;
 use Hyperf\Metric\MetricSetter;
 use Hyperf\Utils\Coroutine;
+use Hyperf\Utils\System;
 use Psr\Container\ContainerInterface;
 use Swoole\Server as SwooleServer;
 
@@ -90,21 +92,27 @@ class OnMetricFactoryReady implements ListenerInterface
             'metric_process_memory_peak_usage'
         );
 
-        $swooleServer = null;
+        $serverStatsFactory = null;
 
-        if (! MetricFactoryPicker::$isCommand && $this->container->has(SwooleServer::class) && $server = $this->container->get(SwooleServer::class)) {
-            if ($server instanceof SwooleServer) {
-                $swooleServer = $server;
+        if (! MetricFactoryPicker::$isCommand) {
+            if ($this->container->has(SwooleServer::class) && $server = $this->container->get(SwooleServer::class)) {
+                if ($server instanceof SwooleServer) {
+                    $serverStatsFactory = fn (): array => $server->stats();
+                }
+            }
+
+            if (! $serverStatsFactory) {
+                $serverStatsFactory = fn (): array => $this->container->get(CoroutineServerStats::class)->toArray();
             }
         }
 
         $timerInterval = $this->config->get('metric.default_metric_interval', 5);
-        $timerId = $this->timer->tick($timerInterval, function () use ($metrics, $swooleServer) {
+        $timerId = $this->timer->tick($timerInterval, function () use ($metrics, $serverStatsFactory) {
             $this->trySet('', $metrics, Co::stats());
             $this->trySet('timer_', $metrics, Timer::stats());
 
-            if ($swooleServer) {
-                $this->trySet('', $metrics, $swooleServer->stats());
+            if ($serverStatsFactory) {
+                $this->trySet('', $metrics, $serverStatsFactory());
             }
 
             if (class_exists('Swoole\Timer')) {
@@ -112,7 +120,7 @@ class OnMetricFactoryReady implements ListenerInterface
             }
 
             $load = sys_getloadavg();
-            $metrics['sys_load']->set(round($load[0] / swoole_cpu_num(), 2));
+            $metrics['sys_load']->set(round($load[0] / System::getCpuCoresNum(), 2));
             $metrics['metric_process_memory_usage']->set(memory_get_usage());
             $metrics['metric_process_memory_peak_usage']->set(memory_get_peak_usage());
         });