+++ src/metric/src/Middleware/MetricMiddleware.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 namespace Hyperf\Metric\Middleware;
 
 use Hyperf\HttpServer\Router\Dispatched;
+use Hyperf\Metric\CoroutineServerStats;
 use Hyperf\Metric\Timer;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
@@ -20,6 +21,10 @@ use Psr\Http\Server\RequestHandlerInterface;
 
 class MetricMiddleware implements MiddlewareInterface
 {
+    public function __construct(protected CoroutineServerStats $stats)
+    {
+    }
+
     /**
      * Process an incoming server request.
      * Processes an incoming server request in order to produce a response.
@@ -34,9 +39,20 @@ class MetricMiddleware implements MiddlewareInterface
             'request_method' => $request->getMethod(),
         ];
         $timer = new Timer('http_requests', $labels);
+
+        ++$this->stats->accept_count;
+        ++$this->stats->request_count;
+        ++$this->stats->connection_num;
+
         $response = $handler->handle($request);
+
         $labels['request_status'] = (string) $response->getStatusCode();
         $timer->end($labels);
+
+        ++$this->stats->close_count;
+        ++$this->stats->response_count;
+        --$this->stats->connection_num;
+
         return $response;
     }